chainlit/src/requirements.txt
langchain==0.1.9
openai==1.30.3
langchain_core==0.1.26
langchain_community==0.0.21
pydantic==2.5.2
redis==5.0.1
tiktoken==0.6.0
python-dotenv==1.0.1
chainlit==1.1.202
chromadb==0.4.24
sentence_transformers==2.4.0
torch==2.2.0
flashrank==0.1.66
transformers==4.37.2
accelerate==0.27.2
bitsandbytes==0.42.0
faiss-cpu==1.7.4
openpyxl==3.1.2
pypdf==4.0.2
docx2txt==0.8
# traceloop-sdk==0.14.1
unstructured==0.12.4
unstructured-client==0.18.0
unstructured-inference==0.7.24
unstructured.pytesseract==0.3.12
pytesseract==0.3.10
pdf2image==1.17.0
pdfminer.six==20221105
pdfplumber==0.10.4
pikepdf==8.13.0
opencv-python==4.9.0.80
Pillow==10.0.1
pillow_heif==0.15.0
poppler-utils==0.1.0
langchainhub==0.1.15
duckduckgo-search==5.1.0
google-search-results==2.4.2
rank_bm25==0.2.2
numexpr==2.9.0
asyncpg==0.29.0
SQLAlchemy==2.0.23
boto3==1.34.122
azure-identity==1.16.0
azure-storage-file-datalake==12.15.0
# chromadb-data-pipes==0.0.8
# tqdm
# numpy
# scipy
# huggingface-hub
# Pillow
# transformers